Fix subprocess error, fix tagged primer check error, add probe design.
Subprocess was giving too many files open error.  Fixed with os.close() following the call to subprocess.

The comparison of tagged primers to appropriate values was incorrect because of (1) strings were used instead of floats in the dict and (2) the comparison operator was inverted.

Added option for design of DNA probes.