Changing the _best() method to accommodate a _good() method.
Previously, the _best() method just returned the best, tagged primer of the group, while also indicating if the best primer met minimum criteria.  That was sort of silly, since users may be interested in those primers that are good, in addition to that primer which is best.

So, I added a _good() method that returns a list of tagged primers meeting minimum criteria (in terms of self-, end-, and hairpins). From this list, the "best" primer is then chosen (in _best()), negating the need, previously, for the tagged_best_okay attribute.