fix several tag-name issues and add example file
apparently, during the transition from primer3 versions, i held over some of
the older tag names for tags that are not longer functioning.  this is
problematic, because primer3 doesn't complain when improper tags are specified,
rather it keeps running (unless one uses -strict_tags on the CLI).

This update fixes that problem and includes an example file with the tags
in the code.  check with:

primer3_core -strict_tags < example.p3